% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balk.r
\name{binomial_naive_bayes}
\alias{binomial_naive_bayes}
\title{\code{binomial_naive_bayes}}
\usage{
binomial_naive_bayes(x, y, prior = NULL, laplace = 1, ...)
}
\arguments{
\item{x}{a matrix with numeric: 0,1, up to binomial_n columns}

\item{y}{a factor or character or logical vector}

\item{prior}{a vector of prior probabilities}

\item{laplace}{a numeric value for Laplace smoothing}

\item{...}{additional arguments}
}
\value{
return a binomial_naive_bayes object
}
\description{
\code{binomial_naive_bayes} is an implementation of the binomial naive bayes algorithm.
modified from bernoulli_naive_bayes function in the naivebayes package
}
