% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guided_helpers.r
\name{cal_fn}
\alias{cal_fn}
\title{\code{cal_fn}}
\usage{
cal_fn(pattern, goi, target)
}
\arguments{
\item{pattern}{the pattern from \code{generate_pattern}}

\item{goi}{the group of interest (names of isolates)}

\item{target}{the target sequence(s)}
}
\value{
proportion: no. false negative/number of isolates
}
\description{
\code{cal_fn} is used to check if the proportion of false negative
fastas and metas are compatible.
}
