% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guided_helpers.r
\name{cal_fp}
\alias{cal_fp}
\title{\code{cal_fp}}
\usage{
cal_fp(pattern, goi, target)
}
\arguments{
\item{pattern}{the pattern from \code{generate_pattern}}

\item{goi}{the group of interest (names of isolates)}

\item{target}{the target sequence(s)}
}
\value{
proportion: no. false positive/number of isolates
}
\description{
\code{cal_fp} is used to check if the proportion of false positive
fastas and metas are compatible.
}
