% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{cal_met_snp}
\alias{cal_met_snp}
\title{\code{cal_met_snp}}
\usage{
cal_met_snp(position, metric, seqc, ...)
}
\arguments{
\item{position}{position to check}

\item{metric}{either `simpson` or `percent`}

\item{seqc}{list of sequences, either passed directly from
\code{process_allele} or \code{read_fasta} or equivalence}

\item{...}{other parameters as needed}
}
\value{
return the value at that position,
as well as base pattern for next iteration.
}
\description{
\code{cal_met_snp} is used to calculate the metric at each position
}
\keyword{internal}
