% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{calculate.percent}
\alias{calculate.percent}
\title{\code{calculate.percent} is used to calculate dissimilarity index.}
\usage{
calculate.percent(pattern, goi)
}
\arguments{
\item{pattern}{list of sequences}

\item{goi}{group of interest}
}
\value{
Will return fastaDNA.
}
\description{
\code{calculate.percent} is used to calculate dissimilarity index.
}
