% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addn_search.R
\name{calculate_mcc}
\alias{calculate_mcc}
\title{\code{calculate_mcc}}
\usage{
calculate_mcc(pattern, goi, MUST_HAVE_TARGET = TRUE)
}
\arguments{
\item{pattern}{the SNP profile for each samples}

\item{goi}{the samples belonging to the group of interest}

\item{MUST_HAVE_TARGET}{whether to force the profile to have at least 1 target profile
(the profile containing the most goi)}
}
\value{
the MCC score
}
\description{
\code{calculate_mcc} is used to calculate the MCC score given the SNP profile.
}
