% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addn_search.R
\name{calculate_mcc_multi}
\alias{calculate_mcc_multi}
\title{\code{calculate_mcc_multi}}
\usage{
calculate_mcc_multi(pattern, meta, target = "target", priority = NULL)
}
\arguments{
\item{pattern}{the SNP profile for each samples}

\item{meta}{A data.table containing the meta data}

\item{target}{the column name of the target in the meta data, default to target}

\item{priority}{A data.table of the targets and priority,
either supplied by user, or by default generated by \code{generate_prioritisation}.}
}
\value{
multiclass-MCC score
}
\description{
\code{calculate_mcc_multi} Calculate the multi-class MCC score for the SNPs.
It assigns each SNP profile to a class, based on the majority of the samples having the profile,
targets with the less samples are prioritised first, breaking ties by alphabetical order.
}
