% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{calculate_percent}
\alias{calculate_percent}
\title{\code{calculate_percent}}
\usage{
calculate_percent(pattern, goi)
}
\arguments{
\item{pattern}{list of sequences' pattern (profile)}

\item{goi}{group of interest}
}
\value{
Will return the dissimilarity index of the list of patterns.
}
\description{
\code{calculate_percent} is used to calculate dissimilarity index,
proportion of isolates not in goi that have been discriminated against.
1 being all and 0 being none.
}
