% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{calculate_simpson}
\alias{calculate_simpson}
\title{\code{calculate_simpson}}
\usage{
calculate_simpson(pattern)
}
\arguments{
\item{pattern}{list of sequences}
}
\value{
Will return the Simpson's index of the list of patterns.
}
\description{
\code{calculate_simpson} is used to calculate Simpson's index.
Which is in the range of 0-1, where the greater the value,
the more diverse the population.
}
