% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addn_search.R
\name{calculate_simpson_by_group}
\alias{calculate_simpson_by_group}
\title{\code{calculate_simpson_by_group}}
\usage{
calculate_simpson_by_group(pattern, meta, target)
}
\arguments{
\item{pattern}{list of sequences' pattern (profile)}

\item{meta}{the metadata}

\item{target}{the target column name}
}
\value{
Will return the Simpson's index of the list of patterns.
}
\description{
\code{calculate_simpson_by_group} is used to calculate Simpson's index.
Which is in the range of 0-1, where the greater the value,
the more diverse the population.
}
