% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{calculate_state}
\alias{calculate_state}
\title{\code{calculate_state}}
\usage{
calculate_state(pattern)
}
\arguments{
\item{pattern}{list of sequences' pattern (profile)}
}
\value{
number of states
}
\description{
\code{calculate_state} calculate the number of states given the SNP(s)
}
