% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{calculate_variant_within_group}
\alias{calculate_variant_within_group}
\title{\code{identify_group_variant_breakdown}}
\usage{
calculate_variant_within_group(pattern, meta, target, get_count = FALSE)
}
\arguments{
\item{pattern}{list of sequences' pattern (profile)}

\item{meta}{metadata of the sequences}

\item{target}{column name of the target group}

\item{get_count}{whether to return the count of samples rather than the raw number, default to FALSE.}
}
\value{
Will return the Simpson's index of the list of patterns.
}
\description{
\code{calculate_variant_within_group} is used to identify proportion of different samples
having the same profile.
}
