% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{check_fasta_meta_mapping}
\alias{check_fasta_meta_mapping}
\title{\code{check_fasta_meta_mapping}}
\usage{
check_fasta_meta_mapping(fasta, meta)
}
\arguments{
\item{fasta}{the fasta read into memory to join}
}
\description{
\code{check_fasta_meta_mapping} is used to check if
fastas and metas are compatible.
}
