% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addn_search.R
\name{check_meta_target}
\alias{check_meta_target}
\title{\code{check_meta_target}}
\usage{
check_meta_target(list_of_parameters)
}
\arguments{
\item{list_of_parameters}{is a list of parameter passed
to functions that will perform the calculation}
}
\value{
TRUE if the parameters exists, else FALSE
}
\description{
\code{check_meta_target} is used to check if parameters needed by
\code{calculate_mcc_multi} and \code{simpson_by_group} are all present.
}
