% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{check_percent}
\alias{check_percent}
\title{\code{check_percent}}
\usage{
check_percent(list_of_parameters)
}
\arguments{
\item{list_of_parameters}{is a list of parameter passed
to functions that will perform the calculation}
}
\value{
TRUE if goi exists, else FALSE
}
\description{
\code{check_percent} is used to check if parameters needed by
\code{calculate_percent} are all present.
}
