% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balk.r
\name{coef.binomial_naive_bayes}
\alias{coef.binomial_naive_bayes}
\title{\code{coef.binomial_naive_bayes}}
\usage{
\method{coef}{binomial_naive_bayes}(object, ...)
}
\arguments{
\item{object}{a binomial_naive_bayes object}

\item{...}{additional arguments}
}
\value{
return a data frame of coefficients
}
\description{
\code{coef.binomial_naive_bayes} is an implementation of the coef method for the binomial naive bayes algorithm.
modified from bernoulli_naive_bayes function in the naivebayes package
}
