% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_from_reads.R
\name{combine_fastq_search_result}
\alias{combine_fastq_search_result}
\title{\code{combine_fastq_search_result}}
\usage{
combine_fastq_search_result(
  results,
  search_table,
  previous_result = NULL,
  bp = MulticoreParam()
)
}
\arguments{
\item{results}{the result (fastq_search_result) from \code{search_from_fastq_reads} to combine.}

\item{search_table}{a dataframe with the following columns:
- "id","type","sequence","strand","result","extra","match_ref_seq"}

\item{previous_result}{the result (fastq_search_result) to append to}

\item{bp}{BiocParallel backend to use for parallelization}
}
\value{
will return a dataframe containing: -
`sequence`, `search_id`, `reads`, `raw_match`, `mean_qualities`, `indexes`, `id`, `type`,
`strand`, `result`, `extra`, `match_ref_seq`, `n_reads`
}
\description{
\code{combine_fastq_search_result} combines the search results from \code{search_from_fastq_reads}
}
