% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_from_reads.R
\name{combine_search_string_result_from_list}
\alias{combine_search_string_result_from_list}
\title{\code{combine_search_string_result_from_list}}
\usage{
combine_search_string_result_from_list(
  results,
  search_table,
  append_to_current_result = data.frame(),
  bp = MulticoreParam()
)
}
\arguments{
\item{results}{the dataframes from \code{search_from_fastq_reads} to combine.}

\item{search_table}{a dataframe with the following columns:
- "id","type","sequence","strand","result","extra","match_ref_seq"}

\item{append_to_current_result}{the dataframe of previous result to append to}

\item{bp}{BiocParallel backend to use for parallelization}
}
\value{
will return a dataframe containing: -
`sequence`, `search_id`, `reads`, `raw_match`, `mean_qualities`, `indexes`, `id`, `type`,
`strand`, `result`, `extra`, `match_ref_seq`, `n_reads`
}
\description{
\code{combine_search_string_result_from_list} combines the search results from \code{search_from_fastq_reads}
}
