% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_from_reads.R
\name{estimate_coverage}
\alias{estimate_coverage}
\title{\code{estimate_coverage}}
\usage{
estimate_coverage(read_lengths, genome_size)
}
\arguments{
\item{read_lengths}{the lengths of the reads}

\item{genome_size}{the genome size}
}
\value{
will return an estimated average coverage
}
\description{
\code{estimate_coverage}
\code{estimate_coverage} estimate the average coverage by comparing number of bases from reads to genome size
}
