% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{find_optimised_snps}
\alias{find_optimised_snps}
\title{\code{find_optimised_snps} is used to find optimised SNPs set.}
\usage{
find_optimised_snps(
  seqc,
  criteria = "simpson",
  goi = c(),
  accept_multiallelic = FALSE,
  number_of_result = 1,
  max_depth = 1,
  included_positions = c(),
  excluded_positions = c(),
  iterate_included = FALSE,
  bp = SerialParam()
)
}
\arguments{
\item{seqc}{list of sequences}

\item{criteria}{either `simpson` or `percent`}

\item{goi}{group of interest, if creteria is percent,
must be specified, ignored otherwise}

\item{accept_multiallelic}{whether include positions with > 2 states}

\item{number_of_result}{number of results to return}

\item{max_depth}{maximum depth to go before terminating}

\item{included_positions}{included positions}

\item{excluded_positions}{excluded positions}

\item{iterate_included}{whether to calculate index
at each level of the included SNPs}

\item{bp}{BiocParallel backend}
}
\value{
Will return the resolution-optimised SNPs set, based on the criteria.
}
\description{
\code{find_optimised_snps} is used to find optimised SNPs set.
}
