% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{flag_allele}
\alias{flag_allele}
\title{\code{flag_allele} is used to find out a list of samples
that has been flagged and will not be included in computation
of minimum SNPs.}
\usage{
flag_allele(seqc, bp = BiocParallel::SerialParam())
}
\arguments{
\item{seqc}{a list containing list of nucleotides. To keep it simple,
use provided read.fasta to import the fasta file.}

\item{bp}{is the biocparallel backend, default to serialParam,
most likely sufficient in most scenario}
}
\value{
Will return a list of ignored samples.
}
\description{
\code{flag_allele} is used to find out a list of samples
that has been flagged and will not be included in computation
of minimum SNPs.
}
