% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{full_merge}
\alias{full_merge}
\title{\code{full_merge}}
\usage{
full_merge(
  fasta_1,
  fasta_2,
  meta_1,
  meta_2,
  ref,
  bp = BiocParallel::MulticoreParam(),
  ...
)
}
\arguments{
\item{fasta_1}{fasta read into memory to join}

\item{fasta_2}{fasta read into memory to join}

\item{meta_1}{meta file for `fasta_1` denoting all positions of SNPs
and position in reference genome}

\item{meta_2}{meta file for `fasta_2` denoting all positions of SNPs
and position in reference genome}

\item{ref}{name of the reference genome (needs to be in both fasta files)}
}
\value{
merged fasta and meta
}
\description{
\code{full_merge} is used to merge 2 fasta,
where a position exist only in 1 of the fasta, the fasta without allele
in that positions are given reference genome's allele at that position.
**Doesn't work for large dataset, hence the need for \code{full_merge_1}**
}
