% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_string_generation.R
\name{generate_kmer_search_string}
\alias{generate_kmer_search_string}
\title{\code{generate_kmer_search_string}}
\usage{
generate_kmer_search_string(
  gene_seq,
  k,
  id_prefix = NULL,
  bp = MulticoreParam()
)
}
\arguments{
\item{gene_seq}{sequences to generate k_mers from}

\item{k}{kmer length}

\item{id_prefix}{prefix for the gene id}

\item{bp}{BiocParallel backend to use}
}
\value{
a dataframe containing the search strings
}
\description{
\code{generate_kmer_search_string} generate the search strings
to detect genes' presence
}
