% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_string_generation.R
\name{generate_kmers}
\alias{generate_kmers}
\title{\code{generate_kmers}}
\usage{
generate_kmers(final_string, k)
}
\arguments{
\item{final_string}{the string to generate kmers}

\item{k}{the length of the kmer}
}
\value{
a vector of kmers
}
\description{
\code{generate_kmers} generate the kmer sequences of the given length
}
