% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{generate_pattern}
\alias{generate_pattern}
\title{\code{generate_pattern} is used to generate pattern for calculation.}
\usage{
generate_pattern(seqc, ordered_index = c(), append_to = list())
}
\arguments{
\item{seqc}{list of sequences}

\item{ordered_index}{list of indexes for the pattern in the order}

\item{append_to}{existing patterns to append to}
}
\value{
Will return fastaDNA.
}
\description{
\code{generate_pattern} is used to generate pattern for calculation.
}
