% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addn_search.R
\name{generate_prioritisation}
\alias{generate_prioritisation}
\title{\code{generate_prioritisation}}
\usage{
generate_prioritisation(meta)
}
\arguments{
\item{meta}{A data.table containing the meta data, expect the}
}
\value{
A data.table of the targets in order of priority
}
\description{
\code{generate_prioritisation} create a vector of the targets in order of priority.
Targets with the less samples are prioritised first, breaking ties by alphabetical order.
}
