% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_from_reads.R
\name{get_all_process_methods}
\alias{get_all_process_methods}
\title{\code{get_all_process_methods}}
\usage{
get_all_process_methods(process_name = "")
}
\arguments{
\item{process_name}{name of the metric, "" to return all,
`SNP` or `KMER` are provided as default.}
}
\value{
a list, including the function to process the search sequence result
}
\description{
\code{get_all_process_methods} is used to get the metrics function
and required parameters. Additional metric may be set by
assigning it to `MinSNPs_process_methods` variable.
}
