% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balk.r
\name{get_binomial_tables}
\alias{get_binomial_tables}
\title{\code{get_binomial_tables}}
\usage{
get_binomial_tables(prob1)
}
\arguments{
\item{prob1}{a matrix of probabilities}
}
\value{
return table of probability for binomial naive bayes
}
\description{
\code{get_binomial_tables} is an internal function that returns a table of probability for binomial naive bayes.
modified from bernoulli_naive_bayes function in the naivebayes package
}
