% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_metric_fun}
\alias{get_metric_fun}
\title{\code{get_metric_fun}}
\usage{
get_metric_fun(metric_name)
}
\arguments{
\item{metric_name}{name of the metric, by default percent/simpson}
}
\value{
a list, including the function to calculate the
metric based on a position (`calc`), and function to check for
additional parameters the function need (`args`)
}
\description{
\code{get_metric_fun} is used to get the metrics function
and required parameters. Additional metric may set by
assigning to `MinSNPs_metrics` variable.
}
