% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_positions_to_search}
\alias{get_positions_to_search}
\title{\code{get_positions_to_search}}
\usage{
get_positions_to_search(seqc_len, excluded_pos, traversed, bp)
}
\arguments{
\item{seqc_len}{length of the matrix}

\item{excluded_pos}{vector of excluded positions}

\item{traversed}{vector of positions that is previously selected}

\item{bp}{BiocParallel backend}
}
\value{
Will return a list of positions to search through
}
\description{
\code{get_positions_to_search} is used to identify all the positions
to loop through, used by \code{branch_and_search} and \code{find_optimised_snps}
}
\keyword{internal}
