% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_snps_set}
\alias{get_snps_set}
\title{\code{get_snps_set}}
\usage{
get_snps_set(results, as = "data.frame")
}
\arguments{
\item{results}{output from `find_optimised_snps`}

\item{as}{output format, either `data.frame` or `list`.}
}
\value{
will return either 
1. a dataframe containing SNPs_set (SNP position separated by ",") and score
2. a list containing SNPs_set (SNP position as numeric vector) and score (attr of the list)
}
\description{
\code{get_snps_set} extract the SNP sets from the output of
`find_optimised_snps`.
}
