% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_usual_length}
\alias{get_usual_length}
\title{\code{get_usual_length} is used to find out
the length of the sequences (W/O deletion).}
\usage{
get_usual_length(seqc, bp = SerialParam())
}
\arguments{
\item{seqc}{a list containing list of nucleotides. To keep it simple,
use provided read_fasta to import the fasta file.}

\item{bp}{is the biocparallel backend, default to serialParam,
most likely sufficient in most scenario}
}
\value{
Will return the length of most of the samples,
the higher number is taken as tie breaker
}
\description{
\code{get_usual_length} is used to find out
the length of the sequences (W/O deletion).
}
\keyword{internal}
