% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_string_generation.R
\name{identify_overlaps}
\alias{identify_overlaps}
\title{\code{identify_overlaps}}
\usage{
identify_overlaps(position_reference, genome_position, prev, after)
}
\arguments{
\item{position_reference}{the mapping of position
in SNP matrix to reference genome}

\item{genome_position}{the position of the SNP in the reference genome}

\item{prev}{number of bases before the SNP included in the search string}

\item{after}{number of bases after the SNP included in the search string}
}
\value{
a list containing 2 dataframes listing the bystander SNPs
in the flanking sequence before and after the SNPs
}
\description{
\code{identify_overlaps} identify the overlapping SNPs in the sequences
}
