% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_from_reads.R
\name{infer_from_combined}
\alias{infer_from_combined}
\title{\code{infer_from_combined}}
\usage{
infer_from_combined(combined_result, search_table, ...)
}
\arguments{
\item{combined_result}{the combined result from \code{combine_search_string_result}}

\item{search_table}{a dataframe with the following columns:
- "id","type","sequence","strand","result","extra","match_ref_seq"}
}
\value{
a dataframe containing the following columns:
- type, rank, result, reads_count, proportion_matched, pass_filter
}
\description{
\code{infer_from_combined} infers the results (presence/absense of genes & CC) from the combined result
}
