% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{iterate_merge}
\alias{iterate_merge}
\title{\code{iterate_merge}}
\usage{
iterate_merge(
  fastas,
  metas,
  ref,
  method = "full",
  bp = BiocParallel::SerialParam(),
  ...
)
}
\arguments{
\item{fastas}{list of fastas read into memory to join}

\item{metas}{list of metas read into memory to join}

\item{ref}{name of the reference genome (needs to be in both fasta files)}

\item{method}{how to join the 2 fasta, currently supported methods are:
inner, full}
}
\value{
Will return a list containing a merged FASTA and a meta.
}
\description{
\code{iterate_merge} is used to combine > 2 fastas iteratively.
}
