% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{iterate_through}
\alias{iterate_through}
\title{\code{iterate_through}}
\usage{
iterate_through(metric, seqc, bp = MulticoreParam(), ...)
}
\arguments{
\item{metric}{either `simpson` or `percent`}

\item{seqc}{list of sequences, either passed directly from
\code{process_allele} or \code{read_fasta} or equivalence}

\item{bp}{BiocParallel backend.
Rule of thumbs: use MulticoreParam(workers = ncpus - 2)}

\item{...}{other parameters as needed}
}
\value{
return a dataframe containing the position and result.
}
\description{
\code{iterate_through} is used to calculate the metric at each position
}
