% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addn_search.R
\name{map_profile_to_target}
\alias{map_profile_to_target}
\title{\code{map_profile_to_target}}
\usage{
map_profile_to_target(meta, patterns, priority)
}
\arguments{
\item{meta}{A data.table containing the meta data}

\item{patterns}{A list of the patterns from \code{generate_pattern}}

\item{priority}{A data.table of the targets and priority, either generated by \code{generate_prioritisation} or supplied by user}
}
\value{
A vector of the targets in order of priority
}
\description{
\code{map_profile_to_target} creates a mapping of the profile to the target, breaking the ties by the priority.
}
