% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{match_count}
\alias{match_count}
\title{\code{match_count}}
\usage{
match_count(target, search_from)
}
\arguments{
\item{target}{the search target}

\item{search_from}{the sequence to search from}
}
\value{
number of matches
}
\description{
\code{match_count} return the number of matches of the target string in the
given sequence
}
