% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addn_search.R
\name{mcc_calculation}
\alias{mcc_calculation}
\title{\code{mcc_calculation}}
\usage{
mcc_calculation(
  result_with_truth,
  is_multi = TRUE,
  return_all_intermediate = FALSE
)
}
\arguments{
\item{result_with_truth}{the dataframe containing the truth and predicted target}

\item{is_multi}{Whether to use MCC-multi or MCC}

\item{return_all_intermediate}{whether to return all intermediate values, only possible for binary class}
}
\value{
Will return the mcc score
}
\description{
\code{mcc_calculation} calculate the MCC score given the truth and predicted target.
}
