% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_parsed}
\alias{output_parsed}
\title{\code{output_parsed}}
\usage{
output_parsed(parsed_view, seqc_obj, result, file_name)
}
\arguments{
\item{parsed_view}{the result from \code{get_metric_fun}, passed from
\code{output_result}}

\item{seqc_obj}{the sequences object.}

\item{result}{is the result from \code{find_optimised_snps}}

\item{file_name}{file name, either csv or "" for terminal output}
}
\value{
NULL, result either printed or saved as csv.
}
\description{
\code{output_parsed} is subfunction used by \code{output_result}
to present the result and save the result as csv.
}
\keyword{internal}
