% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_result}
\alias{output_result}
\title{\code{output_result}}
\usage{
output_result(result, view = "", ...)
}
\arguments{
\item{result}{is the result from \code{find_optimised_snps}}

\item{view}{how to present the output,
"csv" or "tsv" will be saved as a file.
Otherwise, printed to console.}

\item{...}{if view is "tsv" or "csv", file name can be passed, e.g.,
file_name = "result.tsv", otherwise, file is saved as <timestamp>.tsv.}
}
\value{
NULL, result either printed or saved as tsv.
}
\description{
\code{output_result} is used to present the result
and save the result as tsv.
}
