% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{output_to_files}
\alias{output_to_files}
\title{\code{output_to_files}}
\usage{
output_to_files(merged_result, filename = "merged")
}
\arguments{
\item{merged_result}{a list containing the merged fasta and meta.}

\item{filename}{filename to write to, will output
<filename>.fasta and <filename>.csv.}
}
\value{
NULL, files written to filesystem
}
\description{
\code{output_to_files} is write the result to files.
}
