% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balk.r
\name{predict.binomial_naive_bayes}
\alias{predict.binomial_naive_bayes}
\title{\code{predict.binomial_naive_bayes}}
\usage{
\method{predict}{binomial_naive_bayes}(object, newdata = NULL, type = c("class", "prob"), ...)
}
\arguments{
\item{object}{a binomial_naive_bayes object}

\item{newdata}{a matrix with numeric: 0,1,up to binomial_n columns}

\item{type}{a character string specifying the type of output: "class" or "prob"}

\item{...}{additional arguments}
}
\value{
return a factor or matrix of class probabilities
}
\description{
\code{predict.binomial_naive_bayes} is an implementation of the predict method for the binomial naive bayes algorithm.
modified from bernoulli_naive_bayes function in the naivebayes package
}
