% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balk_helper.r
\name{predict_balk}
\alias{predict_balk}
\title{\code{predict_balk}}
\usage{
predict_balk(object, newdata = NULL, snp_id = NULL, type = "prob")
}
\arguments{
\item{object}{The classifier object}

\item{newdata}{A list of sequences}

\item{snp_id}{A vector of SNP IDs}

\item{type}{The type of prediction, either "prob" or "class"}
}
\value{
A vector of either the class or the probability of the class
}
\description{
\code{predict_balk} is a function that predicts the class of new data using a binomial naive bayes classifier
}
