% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balk.r
\name{print.binomial_naive_bayes}
\alias{print.binomial_naive_bayes}
\title{\code{print.binomial_naive_bayes}}
\usage{
\method{print}{binomial_naive_bayes}(x, ...)
}
\arguments{
\item{x}{a binomial_naive_bayes object}

\item{...}{additional arguments}
}
\value{
return a binomial_naive_bayes object
}
\description{
\code{print.binomial_naive_bayes} is an implementation of the print method for the binomial naive bayes algorithm.
modified from bernoulli_naive_bayes function in the naivebayes package
}
