% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_allele}
\alias{process_allele}
\title{\code{process_allele} is used to returned the processed allelic
profiles.}
\usage{
process_allele(seqc, bp = BiocParallel::SerialParam())
}
\arguments{
\item{seqc}{a list containing list of nucleotides. To keep it simple,
use provided read.fasta to import the fasta file.}

\item{bp}{is the biocparallel backend, default to serialParam,
most likely sufficient in most scenario}
}
\value{
Will return the processed allelic profiles.
}
\description{
\code{process_allele} is used to returned the processed allelic
profiles.
}
