% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_from_reads.R
\name{process_kmer_result}
\alias{process_kmer_result}
\title{\code{process_kmer_result}}
\usage{
process_kmer_result(partial_result, search_table, min_match_per_read = 1, ...)
}
\arguments{
\item{partial_result}{the result from \code{infer_from_combined} with only KMER}

\item{search_table}{a dataframe with the following columns:
- "id","type","sequence","strand","result","extra","match_ref_seq"}

\item{min_match_per_read}{the minimum number of kmer matches in a read, discarding reads with less than this number}

\item{...}{ignored}
}
\value{
a dataframe containing the following columns:
- type, rank, result, reads_count, proportion_matched, pass_filter
}
\description{
\code{process_kmer_result} processes the KMER result from \code{infer_from_combined}
}
