% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{process_result_file}
\alias{process_result_file}
\title{\code{process_result_file}}
\usage{
process_result_file(result_filepath)
}
\arguments{
\item{result_filepath}{is the path of the saved output file.}
}
\value{
will return a list containing SNPs_set (SNP position as numeric vector).
}
\description{
\code{process_result_file} extract the SNP sets from the saved output file.
}
