% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_from_reads.R
\name{process_snp_result}
\alias{process_snp_result}
\title{\code{process_snp_result}}
\usage{
process_snp_result(
  partial_result,
  search_table,
  count_measure = "n_reads",
  ...
)
}
\arguments{
\item{partial_result}{the result from \code{infer_from_combined} with only SNP}

\item{search_table}{a dataframe with the following columns:
- "id","type","sequence","strand","result","extra","match_ref_seq"}

\item{count_measure}{the column name of the count measure to use for removing the conflicts}

\item{...}{ignored}
}
\value{
a list containing:
- result: a dataframe containing the following columns:
 - type, rank, result, reads_count, proportion_matched, pass_filter, proportion_scheme_found, details
- snps_found: a vector containing the SNPs ID that have been identified without conflict
- proportion_snps_found: the proportion of SNPs found without conflict
}
\description{
\code{process_snp_result} processes the SNP result from \code{infer_from_combined}
}
