% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addn_search.R
\name{profile_to_group_result}
\alias{profile_to_group_result}
\title{\code{profile_to_group_result}}
\usage{
profile_to_group_result(patterns, profile_target)
}
\arguments{
\item{patterns}{the SNP profile for each samples}

\item{profile_target}{the profile target - should be from samples previously seen, generate with \code{map_profile_to_target}}
}
\value{
Will return the result, given the SNP profile.
}
\description{
\code{profile_to_group_result} given profile target, return the result
}
