% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{read.fasta}
\alias{read.fasta}
\title{\code{read.fasta} is used to read fasta file, implementation
similar to seqinr, but much simpler and allow for spaces in sample name.}
\usage{
read.fasta(file, force_to_lower = TRUE)
}
\arguments{
\item{file}{file path}

\item{force_to_lower}{whether to transform sequences
to lower case, default to TRUE}
}
\value{
Will return fastaDNA.
}
\description{
\code{read.fasta} is used to read fasta file, implementation
similar to seqinr, but much simpler and allow for spaces in sample name.
}
