% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{read_fasta}
\alias{read_fasta}
\title{\code{read_fasta}}
\usage{
read_fasta(file, force_to_upper = TRUE)
}
\arguments{
\item{file}{file path}

\item{force_to_upper}{whether to transform sequences
to upper case, default to TRUE}
}
\value{
Will return list of named character vectors.
}
\description{
\code{read_fasta} is used to read fasta file, implementation
similar to seqinr, but much simpler and allow for spaces in sample name.
}
