% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{remove_dup_isolate}
\alias{remove_dup_isolate}
\title{\code{remove_dup_isolate}}
\usage{
remove_dup_isolate(seqc)
}
\arguments{
\item{seqc}{a list containing list of nucleotides. To keep it simple,
use provided read_fasta to import the fasta file.}
}
\value{
return the list of samples where only the first instance of
the isolate with the duplicate name is kept
}
\description{
\code{remove_dup_isolate} is used to remove isolate with the same name.
}
\keyword{internal}
