% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_from_reads.R
\name{remove_snp_conflict}
\alias{remove_snp_conflict}
\title{\code{remove_snp_conflic}}
\usage{
remove_snp_conflict(result, count_measure = "n_reads")
}
\arguments{
\item{result}{the result from \code{infer_from_combined}}

\item{count_measure}{the column name of the count measure to use for removing the conflicts}
}
\value{
a dataframe containing the same columns as the input result with row containing conflicts removed
}
\description{
\code{remove_snp_conflic} removes the reads with SNPs conflicts from the result
}
