% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{resolve_IUPAC_missing}
\alias{resolve_IUPAC_missing}
\title{\code{resolve_IUPAC_missing}}
\usage{
resolve_IUPAC_missing(
  seqc,
  log_operation = TRUE,
  log_file = "replace.log",
  max_ambiguity = -1,
  replace_method = "random",
  N_is_any_base = FALSE,
  output_progress = TRUE
)
}
\arguments{
\item{seqc}{the sequences to be processed}

\item{log_operation}{whether to log the operation}

\item{log_file}{log file to write the operations}

\item{max_ambiguity}{proportion of ambiguity codes to tolerate,
-1 = ignore. Default to -1}

\item{replace_method}{how to substitute the ambiguity codes,
current supported methods:random and most_common, default to "random".}

\item{N_is_any_base}{whether to treat N as any base or substitute it
with one of the alleles found at the position.}

\item{output_progress}{whether to output progress}
}
\value{
Will return the processed sequences.
}
\description{
\code{resolve_IUPAC_missing} is used to replace the
ambiguity codes found in the sequences.
}
