% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reverse_complement}
\alias{reverse_complement}
\title{\code{reverse_complement}}
\usage{
reverse_complement(seq)
}
\arguments{
\item{seq}{the sequence to reverse complement}
}
\value{
reverse complemented sequence
}
\description{
\code{reverse_complement} returns the reverse complement of the
given sequence
}
