% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{scramble_sequence}
\alias{scramble_sequence}
\title{\code{scramble_sequence}}
\usage{
scramble_sequence(seqc, seed)
}
\arguments{
\item{seqc}{the sequence to scramble}

\item{seed}{the seed to use for scrambling}
}
\value{
a named list, containing the scambled sequence and the new positions
}
\description{
\code{scramble_sequence} scramble the orthologous matrix based on a seed
}
