% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_from_reads.R
\name{sequence_reads_match_count}
\alias{sequence_reads_match_count}
\title{\code{sequence_reads_match_count}}
\usage{
sequence_reads_match_count(search_sequence, reads, qualities)
}
\arguments{
\item{search_sequence}{the search sequence to look for where `.` stands for any character.}

\item{reads}{the sequences reads to search for.}

\item{qualities}{the qualities of each bases in the reads.}
}
\value{
will return a list containing for each read: -
count, mean_quality, indexes
}
\description{
\code{sequence_reads_match_count} look for the search sequences in reads and return the matches indexes and mean qualities
}
