% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addn_search.R
\name{summarise_result}
\alias{summarise_result}
\title{\code{summarise_result}}
\usage{
summarise_result(
  snp_sets,
  training_seqs,
  validation_seqs,
  training_metadata,
  validation_metadata,
  priority,
  is_multi = TRUE,
  return_all_intermediate = FALSE,
  is_percent = FALSE
)
}
\arguments{
\item{snp_sets}{the dataframe containing the truth and predicted target}

\item{training_seqs}{the training sequences}

\item{validation_seqs}{the validation sequences}

\item{training_metadata}{the training metadata}

\item{validation_metadata}{the validation metadata}

\item{priority}{the priority of the target, generated by \code{generate_prioritisation}}

\item{is_multi}{Whether to use MCC-multi or MCC}

\item{return_all_intermediate}{whether to return all intermediate values, only possible for binary class}

\item{is_percent}{whether to return result by considering all the profiles having a GOI as target profile}
}
\value{
Will return the summarised result
}
\description{
\code{summarise_result} calculate the MCC score given the SNP sets, training, validation and metadata(s).
}
