% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balk.r
\name{summary.binomial_naive_bayes}
\alias{summary.binomial_naive_bayes}
\title{\code{summary.binomial_naive_bayes}}
\usage{
\method{summary}{binomial_naive_bayes}(object, ...)
}
\arguments{
\item{object}{a binomial_naive_bayes object}

\item{...}{additional arguments}
}
\value{
return a summary of the binomial_naive_bayes object
}
\description{
\code{summary.binomial_naive_bayes} is an implementation of the summary method for the binomial naive bayes algorithm.
modified from bernoulli_naive_bayes function in the naivebayes package
}
