% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balk_helper.r
\name{train_balk}
\alias{train_balk}
\title{\code{train_balk}}
\usage{
train_balk(
  seqc,
  snps_pos,
  meta,
  binomial_n = 1,
  laplace = 1,
  snp_id = NULL,
  prior = NULL,
  fit_prior = FALSE
)
}
\arguments{
\item{seqc}{A list of sequences}

\item{snps_pos}{A vector of SNP positions}

\item{meta}{A data frame containing the metadata, require isolate and target columns}

\item{binomial_n}{The number of classes for the binomial naive bayes, default to 1 - bernoulli, 2 - binomial (support heterozygous SNPs)}

\item{laplace}{The Laplace smoothing parameter}

\item{snp_id}{A vector of SNP IDs, if not provided, it will be inferred from the SNP positions}

\item{prior}{The prior probabilities of the classes}

\item{fit_prior}{Whether to learn class prior probabilities}
}
\value{
A list containing the classifier and the transformation levels
}
\description{
\code{train_balk} is a function that trains a binomial naive bayes classifier for sequence data
}
