% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balk_helper.r
\name{transform_snp}
\alias{transform_snp}
\title{\code{transform_snp}}
\usage{
transform_snp(pat, binomial_n, levels = c(), get = c("levels", "transformed"))
}
\arguments{
\item{pat}{A string of a SNP}

\item{binomial_n}{The number of classes for the binomial naive bayes, default to 1 - bernoulli, 2 - binomial (support heterozygous SNPs)}

\item{levels}{Existing transformation levels, if not provided, it will be inferred from the SNP}

\item{get}{What to return, either "levels" or "transformed", or both}
}
\value{
A vector of either the transformation levels or the transformed SNP or a list containing both
}
\description{
\code{transform_snp} is a function that transforms a SNP into a matrix for binomial naive bayes.
}
