% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{translate_position}
\alias{translate_position}
\title{\code{translate_position}}
\usage{
translate_position(position, positions_table, to = "original")
}
\arguments{
\item{position}{the position to translate}

\item{positions_table}{the table containing the original and scrambled positions}

\item{to}{the direction to translate, either "original" or "scrambled"}
}
\value{
the translated position
}
\description{
\code{translate_position} translate the scambled position in the alignment
to the original position or vice versa
}
