% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addn_search.R
\name{view_mcc}
\alias{view_mcc}
\title{\code{view_mcc}}
\usage{
view_mcc(result, ...)
}
\arguments{
\item{result}{is the result from \code{find_optimised_snps}}

\item{...}{other optional parameters}
}
\value{
formatted result list to be saved or presented.
}
\description{
\code{view_mcc} is used to present the result of selected SNPs set
based on the MCC score
}
