% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addn_search.R
\name{view_mcc_multi}
\alias{view_mcc_multi}
\title{\code{view_mcc_multi}}
\usage{
view_mcc_multi(result, ...)
}
\arguments{
\item{result}{is the result from \code{find_optimised_snps}}

\item{...}{other optional parameters}
}
\value{
formatted result list to be saved or presented.
}
\description{
\code{view_mcc_multi} is used to present the result of selected SNPs set
based on the multi-MCC score
}
