% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{view_simpson}
\alias{view_simpson}
\title{\code{view_simpson}}
\usage{
view_simpson(result, ...)
}
\arguments{
\item{result}{is the result from \code{find_optimised_snps}}

\item{...}{other optional parameters}
}
\value{
formatted result list to be saved or presented.
}
\description{
\code{view_simpson} is used to present the result of selected
SNPs set based on Simpson's Index.
}
