% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{write_fasta}
\alias{write_fasta}
\title{\code{write_fasta}}
\usage{
write_fasta(seqc, filename)
}
\arguments{
\item{seqc}{a list containing list of nucleotides. To keep it simple,
use provided read_fasta to import the fasta file.}

\item{filename}{filename of the output file}
}
\value{
will write the alignments to file
}
\description{
\code{write_fasta} is used to write
the named character vectors to fasta file.
}
