% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{calculate.simpson}
\alias{calculate.simpson}
\title{\code{calculate.simpson} is used to calculate Simpson's index.}
\usage{
calculate.simpson(pattern)
}
\arguments{
\item{pattern}{list of sequences}
}
\value{
Will return the Simpson's index of the list of patterns.
}
\description{
\code{calculate.simpson} is used to calculate Simpson's index.
}
