% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_result}
\alias{output_result}
\title{\code{output_result}}
\usage{
output_result(result, view, ...)
}
\arguments{
\item{result}{is the result from \code{find_optimised_snps}}

\item{view}{how to present the output, "csv" will be saved as a file.
Otherwise, printed to console.}

\item{...}{if view is "csv", file name can be passed, e.g.,
file_name = "result.csv", otherwise, file is saved as <timestamp>.csv.}
}
\value{
NULL, result either printed or saved as csv.
}
\description{
\code{output_result} is used to present the result
and save the result as csv.
}
